+++ src/Convert/SeverityTask.hs
@@ -1,50 +1,59 @@
 {- sv2v
  - Author: Ethan Sifferman <ethan@sifferman.dev>
  -
- - Conversion of standard and elaboration severity tasks `$info`, `$warning`,
- - `$error`, and `$fatal` (20.10 and 20.11).
+ - Conversion of severity system tasks (IEEE 1800-2017 Section 20.10) and
+ - elaboration system tasks (Section 20.11) `$info`, `$warning`, `$error`, and
+ - `$fatal`, which sv2v collectively refers to as "severity tasks".
  -
  - 1. Severity task messages are converted into `$display` tasks.
- - 2. `$fatal` tasks run `$finish` directly after running `$display`.
+ - 2. `$fatal` tasks also run `$finish` directly after running `$display`.
  -}
 
 module Convert.SeverityTask (convert) where
 
 import Data.Char (toUpper)
+import Data.Functor ((<&>))
 
+import Convert.Scoper
 import Convert.Traverse
 import Language.SystemVerilog.AST
 
+type SC = Scoper ()
+
 convert :: [AST] -> [AST]
-convert = map $ traverseDescriptions $ traverseModuleItems convertModuleItem
+convert = map $ traverseDescriptions traverseDescription
 
--- Convert Elaboration Severity Tasks
-convertModuleItem :: ModuleItem -> ModuleItem
-convertModuleItem (ElabTask severity taskArgs) =
-    Initial $ elab severity taskArgs "Elaboration" []
-convertModuleItem other =
-    traverseStmts (traverseNestedStmts convertStmt) other
+traverseDescription :: Description -> Description
+traverseDescription = partScoper return traverseModuleItem return traverseStmt
 
--- Convert Standard Severity Tasks
-convertStmt :: Stmt -> Stmt
-convertStmt (SeverityStmt severity taskArgs) =
-    elab severity taskArgs "[%0t]" [Ident "$time"]
-convertStmt other = other
+-- convert elaboration severity tasks
+traverseModuleItem :: ModuleItem -> SC ModuleItem
+traverseModuleItem (ElabTask severity taskArgs) =
+    elab severity taskArgs "elaboration" [] <&> Initial
+traverseModuleItem other = return other
 
-elab :: Severity -> [Expr] -> String -> [Expr] -> Stmt
-elab severity taskArgs prefixStr prefixArgs =
-    Block Seq "" [] [stmtDisplay, stmtFinish]
+-- convert standard severity tasks
+traverseStmt :: Stmt -> SC Stmt
+traverseStmt (SeverityStmt severity taskArgs) =
+    elab severity taskArgs "%0t" [Ident "$time"]
+traverseStmt other = return other
+
+elab :: Severity -> [Expr] -> String -> [Expr] -> SC Stmt
+elab severity args prefixStr prefixArgs = do
+    scopeName <- hierarchyPathM
+    fileLocation <- sourceLocationM
+    let contextArg = String $ msg scopeName fileLocation
+    let stmtDisplay = call "$display" $ contextArg : prefixArgs ++ displayArgs
+    return $ Block Seq "" [] [stmtDisplay, stmtFinish]
     where
-        stmtDisplay = call "$display" $ msg : prefixArgs ++ args
-        msg = String $ prefixStr ++ ' ' : severityToString severity ++ ':' : trailingSpace
-        trailingSpace = if null args then "" else " "
-        args = if severity /= SeverityFatal || null taskArgs
-            then taskArgs
-            else tail taskArgs
+        msg scope file = severityToString severity ++ " [" ++ prefixStr ++ "] "
+            ++ file ++ " - " ++ scope ++ if null displayArgs then "" else " - "
+        displayArgs = if severity /= SeverityFatal || null args
+            then args
+            else tail args
         stmtFinish = if severity /= SeverityFatal
             then Null
-            else call "$finish" argsFinish
-        argsFinish = if null taskArgs then [] else [head taskArgs]
+            else call "$finish" $ if null args then [] else [head args]
 
 call :: Identifier -> [Expr] -> Stmt
 call func args = Subroutine (Ident func) (Args args [])